51 test 테스트 케이스 추가 및 버그 수정 (#52)
* 테스트 케이스 및 버그 수정

* 테스트 케이스 및 버그 수정2

* 테스트 케이스 및 버그 수정3

* 테스트 케이스 및 버그 수정4

* 테스트 케이스 및 버그 수정5

+++ src/main/domain/organization/joined_organization/service/joined_organization_service.py
@@ -1,10 +1,12 @@
 from fastapi import HTTPException, status
 from sqlalchemy.orm import Session
 
+from common.database import MemberRole
 from common.security.member_DTO import MemberDTO
 from domain.member.entity import Member
 from domain.member.repository import MemberRepository
 from domain.organization.joined_organization.dto import ChangeRoleDto, JoinedOrganizationResponse
+from domain.organization.joined_organization.dto.delete_joined_organization_params import DeleteJoinedOrganizationParams
 from domain.organization.joined_organization.entity import JoinedOrganization
 from domain.organization.joined_organization.repository import JoinedOrganizationRepository
 from domain.organization.organization.dto import OrganizationResponseDto
@@ -24,6 +26,20 @@ class JoinedOrganizationService:
         self.member_repository = member_repository
         self.organization_repository = organization_repository
 
+    async def check_if_owner(self, db:Session, organization_id:int, member_id:int):
+        member = await self.member_repository.find_by_id(db, member_id)
+        if not member:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Member not found")
+        organization = await self.organization_repository.find_by_id(db, organization_id)
+        if not organization:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Organization not found")
+        joined_organization = await self.joined_organization_repository.find_by_member_and_organization(db, member, organization)
+        if not joined_organization:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Member not found")
+        if joined_organization.member_role == MemberRole.OWNER:
+            return True
+        else:
+            return False
     async def change_member_role(
             self,
             db:Session,
@@ -47,16 +63,16 @@ class JoinedOrganizationService:
             db:Session,
             me_dto:MemberDTO
     ):
-        member = await self.member_repository.find_by_id(db, me_dto.member_id)
+        member = await self.member_repository.find_by_id(db, me_dto.id)
         if not member:
             raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Member not found")
-        joined_organizations:list[JoinedOrganization] = self.joined_organization_repository.find_all_by_member(db, member.id)
+        joined_organizations:list[JoinedOrganization] = await self.joined_organization_repository.find_all_by_member(db, member.id)
         organizations = []
         for joined_organization in joined_organizations:
             organization:Organization = joined_organization.organization
             organization_dto = OrganizationResponseDto.model_validate(organization)
             members = []
-            joined_members:list[JoinedOrganization] = organization.joined_organization
+            joined_members:list[JoinedOrganization] = organization.joined_organizations
             for joined_member in joined_members:
                 member:Member = joined_member.member
                 joined_member_dto = JoinedOrganizationResponse.model_validate(joined_member)
@@ -70,9 +86,11 @@ class JoinedOrganizationService:
     async def delete_joined_organization(
             self,
             db:Session,
-            joined_organization_id:int,
+            delete_joined_organization:DeleteJoinedOrganizationParams,
     ):
-        joined_organization = await self.joined_organization_repository.find_by_id(db, joined_organization_id)
+        joined_organization = await self.joined_organization_repository.find_by_id(db, delete_joined_organization.joined_organization_id)
         if not joined_organization:
             raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="JoinedOrganization not found")
+        if joined_organization.organization_id != delete_joined_organization.organization_id:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Organization not matches")
         await self.joined_organization_repository.delete_joined_organization(db, joined_organization)
